Rename and remove some interfaces, fix problems, use generics for server responses types
