+++ frontend/src/common/models/auth/IRegisterUser.ts
@@ -1,4 +1,4 @@
-export interface ISignUpFields {
+export interface IRegisterUser {
   email: string;
   password: string;
   fullName: string;
