--- frontend/src/common/models/auth/IUserInput.ts
@@ -1,5 +1,0 @@
-export interface IUserInput {
-  email: string;
-  password: string;
-  fullName?: string;
-}
