+++ frontend/src/common/models/user/IUser.ts
@@ -3,6 +3,6 @@ export interface IUser {
   email: string;
   fullName: string;
   displayName: string;
-  imageUrl?: string | null | undefined;
-  title?: string | null | undefined;
+  imageUrl?: string | null;
+  title?: string | null;
 }
