--- frontend/src/common/models/user/IUserState.ts
@@ -1,7 +1,0 @@
-import { IUser } from './IUser';
-
-export interface IUserState {
-  data: IUser | null;
-  isLoading: boolean;
-  isAuthorized: boolean;
-}
