+++ frontend/src/components/InputField/styles.module.sass
@@ -1,6 +1,6 @@
 .inputGroup
-  width: '300px'
+  width: 300px
 
 .error
-  font-size: '12px'
-  padding: '2px .5rem'
+  font-size: 12px
+  padding: 2px .5rem
