+++ frontend/src/scenes/SignIn/styles.module.sass
@@ -2,5 +2,5 @@
   height: 100vh
   background-color: var(--auth-form-background)
 
-.signIn-header
+.signInHeader
   padding: 5rem 0 1.5rem 0
